+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
+## [Unreleased]
+
+### Fixed
+- Dynamic scope filtering based on multiple user attribute [RCIAM-1267](https://jira.argo.grnet.gr/browse/RCIAM-1267)
+
 ## [18.0.1-2.13] - 2023-06-07
 
 ### Added